# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIMRS RSUD Dr. H. Chasan Boesoirie Ternate - API Documentation",
...
        "contact": {
            "name": "IT SIMRS RSUD Dr. H. Chasan Boesoirie",
            "url": "https://chasanboesoirie.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/service/medifirst2000",
            "description": "Server SIMRS (Proxy Anti-CORS - Aktif Otomatis)"
        },
        {
            "url": "https://chasanboesoirie.id/service/medifirst2000",
            "description": "Server Langsung (chasanboesoirie.id - Butuh Postman/No-CORS)"
        }
    ],
    "paths": {
...
        "/registrasi/daftar-riwayat-registrasi": {
            "get": {
                "tags": [
                    "Riwayat Pasien (Kunjungan & Registrasi)"
                ],
                "summary": "Daftar Riwayat Kunjungan / Registrasi Pasien",
                "description": "Menampilkan daftar seluruh episode kunjungan/pendaftaran pasien ke rumah sakit (Rawat Jalan, IGD, Rawat Inap). Menampilkan tanggal pendaftaran, nomor registrasi, ruangan/poli tujuan, dokter penanggung jawab, tipe penjamin (BPJS/Umum), tanggal pulang, serta status rawat inap (1: Inap, 0: Jalan/IGD).",
                "operationId": "e0166c1694803e29541b8bdffeb27ea3",
                "parameters": [
                    {
                        "name": "norm",
                        "in": "query",
                        "description": "Nomor Rekam Medis (No CM Pasien)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "0487727"
                        }
                    },
                    {
                        "name": "namaPasien",
                        "in": "query",
                        "description": "Filter nama pasien",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "noReg",
                        "in": "query",
                        "description": "Filter nomor registrasi kunjungan tertentu",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "2605000074"
                        }
                    },
                    {
                        "name": "idRuangan",
                        "in": "query",
                        "description": "Filter ID ruangan/poli",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Riwayat registrasi pasien berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "daftar": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "norec": {
                                                        "type": "string",
                                                        "example": "a1b2c3d4..."
                                                    },
                                                    "tglregistrasi": {
                                                        "type": "string",
                                                        "example": "2026-05-04 02:47:00"
                                                    },
                                                    "nocm": {
                                                        "type": "string",
                                                        "example": "0487727"
                                                    },
                                                    "noregistrasi": {
                                                        "type": "string",
                                                        "example": "2605000074"
                                                    },
                                                    "namapasien": {
                                                        "type": "string",
                                                        "example": "ADITHA BASRA NY"
                                                    },
                                                    "namaruangan": {
                                                        "type": "string",
                                                        "example": "IGD"
                                                    },
                                                    "namadokter": {
                                                        "type": "string",
                                                        "example": "Dr. FAJRIAH A SUMADAYO"
                                                    },
                                                    "kelompokpasien": {
                                                        "type": "string",
                                                        "example": "BPJS"
                                                    },
                                                    "tglpulang": {
                                                        "type": "string",
                                                        "example": "2026-05-04 08:30:00",
                                                        "nullable": true
                                                    },
                                                    "statusinap": {
                                                        "description": "1 = Rawat Inap, 0 = Rawat Jalan / IGD",
                                                        "type": "integer",
                                                        "example": 0
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "ea@epic"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/registrasi/get-detail-registrasi-pasien": {
            "get": {
                "tags": [
                    "Riwayat Pasien (Kunjungan & Registrasi)"
                ],
                "summary": "Detail Kunjungan & Antrian Registrasi Pasien",
                "description": "Mengambil data detail kunjungan meliputi ruangan/kamar, nomor tempat tidur, jam panggil dokter/perawat, dan status antrian periksa.",
                "operationId": "cd30bea9ad595d0693e32c6bc3a5c576",
                "parameters": [
                    {
                        "name": "noregistrasi",
                        "in": "query",
                        "description": "Nomor Registrasi Kunjungan Pasien",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "2605000074"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail data registrasi berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "norec": {
                                            "type": "string",
                                            "example": "apd-12345"
                                        },
                                        "tglregistrasi": {
                                            "type": "string",
                                            "example": "2026-05-04 02:47:00"
                                        },
                                        "namaruangan": {
                                            "type": "string",
                                            "example": "Poli Penyakit Dalam"
                                        },
                                        "namakelas": {
                                            "type": "string",
                                            "example": "Kelas III"
                                        },
                                        "namakamar": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "nobed": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "namadokter": {
                                            "type": "string",
                                            "example": "Dr. FAJRIAH A SUMADAYO"
                                        },
                                        "statusantrian": {
                                            "type": "string",
                                            "example": "SELESAI_DIPERIKSA"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/registrasi/get-antrian-by-nocm-rev": {
            "get": {
                "tags": [
                    "Riwayat Pasien (Kunjungan & Registrasi)"
                ],
                "summary": "Riwayat Antrian & Diagnosa Pasien by No CM",
                "description": "Mengambil data rekam antrian poli dan diagnosa ICD yang pernah ditegakkan pada pasien.",
                "operationId": "03efef0a8eb6e5f9e1eeee9f1c656526",
                "parameters": [
                    {
                        "name": "noCm",
                        "in": "query",
                        "description": "Nomor Rekam Medis (No CM)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "0487727"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar antrian dan diagnosa riwayat pasien"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Token otorisasi SIMRS Ternate. Default token: eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJoaXMuamtuIn0.FhnuCCWQx9SOMwBcLx3NFMPH45-KUwkblNKQY9Debm26PJ7ygt4-1z7oSMrefb0qfJwKtp02kS1O0lupFcMz1Q",
                "name": "X-AUTH-TOKEN",
                "in": "header"
            }
        }
    },
...
}